Remove Helpers/Utils code.
