+++ src/Event.php
@@ -181,7 +181,7 @@ class Event implements PingableInterface
             throw new \BadMethodCallException();
         }
 
-        $amount = !empty($matches[1]) ? word2number($this->splitCamel($matches[1])) : 1;
+        $amount = !empty($matches[1]) ? $this->wordToNumber($this->splitCamel($matches[1])) : 1;
 
         if (!$amount) {
             throw new \BadMethodCallException();
@@ -364,7 +364,7 @@ class Event implements PingableInterface
     public function on($date)
     {
         $date = \date_parse($date);
-        $segments = \array_only($date, \array_flip($this->fieldsPosition));
+        $segments = \array_intersect_key($date, $this->fieldsPosition);
 
         if ($date['year']) {
             $this->skip(function () use ($date) {
@@ -1198,5 +1198,82 @@ class Event implements PingableInterface
         );
 
         return 'WIN' === $osCode;
+    }
+
+    private function wordToNumber($text)
+    {
+        $data = strtr(
+            $text,
+            [
+                'zero' => '0',
+                'a' => '1',
+                'one' => '1',
+                'two' => '2',
+                'three' => '3',
+                'four' => '4',
+                'five' => '5',
+                'six' => '6',
+                'seven' => '7',
+                'eight' => '8',
+                'nine' => '9',
+                'ten' => '10',
+                'eleven' => '11',
+                'twelve' => '12',
+                'thirteen' => '13',
+                'fourteen' => '14',
+                'fifteen' => '15',
+                'sixteen' => '16',
+                'seventeen' => '17',
+                'eighteen' => '18',
+                'nineteen' => '19',
+                'twenty' => '20',
+                'thirty' => '30',
+                'forty' => '40',
+                'fourty' => '40',
+                'fifty' => '50',
+                'sixty' => '60',
+                'seventy' => '70',
+                'eighty' => '80',
+                'ninety' => '90',
+                'hundred' => '100',
+                'thousand' => '1000',
+                'million' => '1000000',
+                'billion' => '1000000000',
+                'and' => '',
+            ]
+        );
+
+        // Coerce all tokens to numbers
+        $parts = \array_map(
+            function ($val) {
+                return (float) $val;
+            },
+            \preg_split('/[\s-]+/', $data)
+        );
+
+        $tmp = null;
+        $sum = 0;
+        $last = null;
+
+        foreach ($parts as $part) {
+            if (null !== $tmp) {
+                if ($tmp > $part) {
+                    if ($last >= 1000) {
+                        $sum += $tmp;
+                        $tmp = $part;
+                    } else {
+                        $tmp += $part;
+                    }
+                } else {
+                    $tmp *= $part;
+                }
+            } else {
+                $tmp = $part;
+            }
+
+            $last = $part;
+        }
+
+        return $sum + $tmp;
     }
 }
